Update adi.R
My first big program in R ! :D
Alright .. getting down to it .. this file was for a very dirty data set. The structure is pretty impeccable (in my opinion). Having said that, it was giving me atrocious accuracy and measures - this had much more to do with the predictor vars I was inputting.